refactor: refactor api client
